+++ src/ArrayTools.jl
@@ -8,7 +8,6 @@ export
     MaybeArrayAxes,
     MaybeArrayAxis,
     RubberIndex,
-    UndefinedType,
     all_match,
     allof,
     anyof,
@@ -22,7 +21,6 @@ export
     common_indices,
     has_standard_indexing,
     noneof,
-    promote_eltype,
     reversemap,
     all_indices,
     same_axes,
@@ -35,24 +33,37 @@ export
     to_axis,
     to_int,
     to_size,
-    to_type,
-    # storage trait
+    # to_type, # FIXME: deprecated, use `as(T,x)` instead
+
+    # Re-exports from TypeUtils:
+    as,
+    as_eltype,
+    convert_eltype,
+    promote_eltype,
+
+    # Storage trait:
     StorageType,
     AnyStorage,
     FlatStorage,
     to_flat_array,
     is_flat_array,
+
     # Fast arrays and indexing trait:
     IndexingType,
     FastIndexing,
     AnyIndexing,
     to_fast_array,
     is_fast_array,
+
     # Macros:
     @assert_same_axes
 
+using TypeUtils
+
 using Base: OneTo, axes1, tail
 import Base: dotview, getindex, setindex!, to_indices
+
+@deprecate to_type(T, x) as(T, x) true
 
 include("traits.jl")
 include("utils.jl")